refactor: moved fields to it's own file for cleaner code
